LayoutTests/platform/ios-simulator/fast/events/platform-wheelevent-paging-x-in-scrolling-page-expected.txt

PASS document.body.scrollTop is window.expectedScrollTop
FAIL document.body.scrollLeft should be 700. Was 0.
